Allow unbalanced designs and multi-group comparisons
Previously, dmrseq required that two group comparisons be balanced
in the number of samples. This commit adds enhancement referenced
in https://github.com/kdkorthauer/dmrseq/issues/2 and allows for
the testing of a covariate with unequal samples in each group. In
addition, this commit adds support for multiple group comparisons,
as referenced in https://github.com/kdkorthauer/dmrseq/issues/6.