Don't require to iterate over chromosomes
Allows to operate on any number of chromosomes at a time (chunk), including all. This commit also contains two bug fixes: (1) when trimming regions, if only one candidate region was found in a given chunk, it would incorrectly return the indices as a list, (2) when including CpGs with missing coverage in some samples, the check of whether methylation values were constant did not properly handle NAs.