Bug fix to properly handle multi-factor conditions in coverage check
This commit fixes issue #6 reported by @LKremer - an error was being
thrown in the check of the number of CpGs with nonzero coverage per
factor group in each region when more than 2 groups were specified.
This was due to improper handling of subsetting the design matrix when
`coeff` was a vector instead of scalar.