Fix bug in constructing permutations for continuous covariate
When constructing unrestricted permutations for large sample
sizes of a factored covariate, there was a check missing for
whether the permutation specified a complete reordering for the
testCovariate, or whether the permutation only specified which
samples are in group one. This resulted in a mispecified design
matrix, leading to a singular error in backsolve. Thanks to
@ND91 for reporting this in issue #11 and to Luli Zou for
reporting and sending a reproducible example over email.