Fix for issue #26
Reported by @MohamedRefaat92 - related to issues #22 and #25,
need to handle the case of two-group factor when no loci with
coverage in one group are present in a given region. Before,
assumed that two groups with coverage would always be present,
and the naming of the columns of the results table was
inconsistent.