Bug fix for issue #37
Bug identified by @zyqfrog10 in the way position indices are tracked
in the internal function bumphunt(). Previously, the order of the
`chrs` vector containing chromosome names did not necessarily match
the order of the names in the cumulative index table when chrM was
included and listed last. This commit
introduces a fix by removing the `as.character` call when building
the cumulative sum table, which was reordering the names alphabetically.
It also adds a check that the order of these two items is identical,
and if not will throw an error to prevent this type of bug from
persisting.